requirements.txt
slackclient==1.1.0
requests==2.18.4
Django==1.11.7
djangorestframework==3.7.3
